requirements.txt
aiofiles==23.2.1
aiogram==2.23.1
aiohttp==3.8.6
aiolimiter==1.1.0
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
APScheduler==3.10.4
async-timeout==4.0.3
attrs==23.2.0
Babel==2.9.1
beautifulsoup4==4.12.3
cachetools==5.3.3
captcha==0.5.0
certifi==2024.2.2
charset-normalizer==3.3.2
class-registry==2.1.2
click==8.1.7
colorama==0.4.6
config==0.5.1
DAWG-Python==0.7.2
docopt==0.6.2
dpath==2.1.6
filters==1.3.2
frozenlist==1.4.1
google==3.0.0
google-search==1.1.1
googlesearch-python==1.2.4
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
libretranslatepy==2.1.1
lxml==5.2.1
magic-filter==1.0.12
multidict==6.0.5
pillow==10.3.0
py-youtube==1.1.7
pydantic==2.5.3
pydantic_core==2.14.6
pyfiglet==1.0.2
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pyTelegramBotAPI==4.17.0
python-dateutil==2.9.0.post0
python-telegram-bot==21.0.1
pytz==2024.1
regex==2023.12.25
requests==2.31.0
six==1.16.0
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.5
telebot==0.0.5
tornado==6.4
translate==3.6.1
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
v==1
yarl==1.9.4
youtube-search==2.1.2
